refactor: rename script to simulate-createProposalWithSolution-txdata
